fix warning for Linux
Former-commit-id: 76b06961f9db3d84f55ca296872f24989e9e293b
Former-commit-id: becdb1257c4c53077c032f038391054124634e43

+++ Frame/SDK/Core/AFRecord.h
@@ -60,10 +60,10 @@ public:
     void SetName(const char* value);
     const char* GetName() const;
 
-    int GetRowCount() const;
+    size_t GetRowCount() const;
 
     void SetColCount(size_t value);
-    int GetColCount() const;
+    size_t GetColCount() const;
 
     bool SetColType(size_t index, int type);
     int GetColType(int col) const;
